ab_transaction

Struct Transaction

pub struct Transaction<'a> {
    pub header: &'a TransactionHeader,
    pub read_slots: &'a [TransactionSlot],
    pub write_slots: &'a [TransactionSlot],
    pub payload: &'a [u128],
    pub seal: &'a [u8],
}

Similar to Transaction, but doesn’t require allow or data ownership.

Can be created with Transaction::as_ref() call.

Fields
header: &'a TransactionHeaderread_slots: &'a [TransactionSlot]

Slots in the form of TransactionSlot that may be read during transaction processing.

These are the only slots that can be used in authorization code.

The code slot of the contract that is being executed and balance of native token are implicitly included and doesn’t need to be specified (see Transaction::read_slots(). Also slots that may also be written to do not need to be repeated in the read slots.

write_slots: &'a [TransactionSlot]

Slots in the form of TransactionSlot that may be written during transaction processing

payload: &'a [u128]seal: &'a [u8]
Implementations

impl Transaction<'_>

pub fn hash(&self) -> TransactionHash

Compute transaction hash.

Note: this computes transaction hash on every call, so worth caching if it is expected to be called often.

pub fn read_slots(&self) -> impl Iterator<Item = TransactionSlot>

Read slots touched by the transaction.

In contrast to read_slots property, this includes implicitly used slots.

pub fn slots(&self) -> impl Iterator<Item = TransactionSlot>

All slots touched by the transaction.

In contrast to read_slots and write_slots properties, this includes implicitly used slots.